+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to
 
 ### Added
 
+- Documentation on `IntegrationError` to clarify that it is not meant to be used
+  directly outside the SDK.
+- New errors that integrations should throw to communicate provider API errors
+  during `validateInvocation` or any step. These will terminate the step, but
+  provide more details to users to help them resolve some kinds of provider API
+  errors.
 - `JobState.addEntity` and `JobState.addRelationship` allow collecting one
   entity/relationship at a time. This supports tight looping and encourages
   quick processing of provider data to minimize memory requirements by
@@ -19,6 +25,12 @@ and this project adheres to
   on-hover documentation rendering in VSCode and other such modern editors. 🤠
   This also allows for communicating to new developers who just jump into the
   code.
+
+### Changed
+
+- Removed an unused error `IntegrationConfigValidationError`.
+- Renamed `IntegrationLocalConfigFieldTypeMismatch` to
+  `IntegrationLocalConfigFieldTypeMismatchError`, to match other error names.
 
 ## 1.1.0 - 2020-05-13